change back to float

+++ detection_embedding/sku.py
@@ -51,7 +51,7 @@ class Sku(Dataset):
             {
             "boxes": group[['x1', 'y1', 'x2', 'y2']].values,
             "labels": np.array([1]*len(group)),
-            "embedding": np.array([[0.1]*512]*len(group), dtype=long)
+            "embedding": np.array([[0.1]*512]*len(group))
             }
             for (image_name, width, height), group
             in groupby]
